+++ controller/commands.rb
@@ -196,230 +196,6 @@ module Controller_Commands
     puts JSON.pretty_generate(Util.opp_view_data(opp))
     
     client.batch_updates(false)
-  end  
-
-  def slack_lookup(slack_params)
-    format_slack_response(find_opportunities(slack_params['text']), slack_params)
-  end
-
-  def find_opportunities(search, limit=nil)
-    limit ||= 4
-    
-    search_esc = Util.escape_sql(search.strip.downcase)
-    
-    from = "FROM #{bigquery.table.query_id} WHERE LOWER(name) LIKE '#{search_esc}' OR links LIKE '%#{search_esc}%' OR emails LIKE '%#{search_esc}%'"
-    counts = bigquery.query("SELECT COUNT(*) total, COUNT(DISTINCT contact) contacts #{from}", '')[0]
-    contacts = bigquery.query("SELECT DISTINCT(contact) contact #{from} LIMIT #{limit}", '').map {|c| c[:contact]}
-    
-    return {count: 0, opportunities: []} if contacts.empty?
-    
-    {
-      count: counts[:total],
-      contacts: counts[:contacts],
-      has_more: (counts[:contacts] > limit),
-      opportunities: client.get_paged_result(OPPORTUNITIES_URL, {contact_id: contacts, expand: client.OPP_EXPAND_VALUES}, 'opportunities_for_contact_ids')
-    }
-  end
-
-  def format_slack_response(results, slack_params)
-    return [{
-    		"type": "section",
-    		"text": {
-    			"type": "mrkdwn",
-    			"text": "No Lever search results found for `#{slack_params['text']}`"
-    		}
-      }] if results[:opportunities].empty?
-    
-    blocks = [
-      {
-    		"type": "section",
-    		"text": {
-    			"type": "mrkdwn",
-    			"text": "Lever search results for `#{slack_params['text']}`#{results[:has_more] ? " (displaying #{results[:opportunities].size} of #{results[:count]} opportunities for #{results[:contacts]} contacts)" : ''}:"
-    		}
-  	  },
-  	  {
-  		"type": "divider"
-  	  }
-  	]
-	
-    results[:opportunities].each{ |opp|
-      opp_data = Util.opp_view_data(opp)
-      blocks += [
-        {
-          "type": "section",
-          "text": {
-            "type": "mrkdwn",
-            "text": "#{opp['archived'].nil? ? '👤 ' : '👻 '}<#{opp['urls']['show']}|*#{opp['name']}* - view on Lever>#{opp['archived'].nil? ? '' : ' [archived]'}" \
-              "#{opp_data['application__posting__text'] ? "\n" + opp_data['application__posting__text'] : ''}" \
-              "\n*Email#{opp['emails'].size > 1 ? 's' : ''}:* #{opp['emails'].join(', ')}" \
-              "#{opp['links'].select{|l| l.include?('linkedin.com')}.any? ? "\n*LinkedIn:* #{opp['links'].select{|l| l.include?('linkedin.com')}.join(', ')}" : ''}" \
-              "\n*Stage:* #{opp['stage']['text']}" \
-              "\n*Last updated:* #{opp_data['lastInteractionAt__datetime']}"
-          }
-        }
-      ]
-    }
-    
-    blocks += [
-      { "type": "divider" },
-      {
-        "type": "context",
-        "elements": [
-          {
-            "type": "mrkdwn",
-            "text": "To search, type `/lever <name, email or url>` - or `/leverme` to show only to yourself."
-          }
-        ]
-      }
-    ]
-    
-    blocks
-  end
-
-  def import_from_bigquery(table_name, test=false)
-    table = bigquery.dataset.table table_name
-    if table.nil? || table.id.nil?
-      log.log("BigQuery table not found: #{bigquery.dataset.dataset_id}.#{table_name}")
-      return
-    end
-    
-    log.log("<Dry-run import - no changes actually applied>") if test
-    
-    log.log("Importing from BigQuery table: #{table.query_id}")
-    rows = bigquery.query("SELECT * FROM #{table.query_id}", '')
-    
-    counts = {
-      rows: 0,
-      errors: 0,
-      existing: 0,
-      new: 0
-    }
-    
-    rows.each { |row|
-      counts[:rows] += 1
-      
-      opp_params = {
-        'email': (Util.get_hash_value_fuzzy(row, 'email') || '').downcase,
-        'linkedin': (Util.get_hash_value_fuzzy(row, 'linkedin') || '').downcase.sub(/\?.+/, ''),
-        'posting': Util.get_hash_value_fuzzy(row, 'posting'),
-        'name': Util.get_hash_value_fuzzy(row, 'name'),
-        'origin': Util.get_hash_value_fuzzy(row, 'origin'),
-        'sources': (Util.get_hash_value_fuzzy(row, 'sources') || '').split(',').map(&:strip),
-        'tags': (Util.get_hash_value_fuzzy(row, 'tags') || '').split(',').map(&:strip),
-        'phones': (Util.get_hash_value_fuzzy(row, 'phone') || '').split(',').map(&:strip),
-        'location': Util.get_hash_value_fuzzy(row, 'location'),
-        'headline': Util.get_hash_value_fuzzy(row, 'headline')
-      }
-      opp_params[:createdAt] = Util.get_hash_value_fuzzy(row, 'createdat')
-      opp_params[:stage] = Util.get_hash_value_fuzzy(row, 'stageid')
-      
-      log.log("Looking for existing opportunity via parameters:\n" + opp_params.select{|k,v| [:email, :linkedin, :posting].include?(k)}.map{|k,v| "- #{k}: #{v}"}.join("\n")) if test
-      
-      if opp_params[:posting].empty?
-        log.log('No job posting ID ("posting") found - unable to process row')
-        counts[:errors] += 1
-        next
-      end
-      
-      opp, is_new = find_or_create_opportunity(opp_params, "Import: #{Time.now.strftime('%Y-%m-%d')} #{table.table_id}", test)
-      
-      if is_new
-        counts[:new] += 1
-      else
-        counts[:existing] += 1
-      end
-
-      fields = row.reject {|k,v|
-        [
-          Util.get_hash_key_fuzzy(row, 'posting'),
-          Util.get_hash_key_fuzzy(row, 'name'),
-          Util.get_hash_key_fuzzy(row, 'email'),
-          Util.get_hash_key_fuzzy(row, 'linkedin')
-        ].include?(k)
-      }.transform_keys {|k|
-        k.to_s.sub(/^coffee_/i, '')
-      }
-      
-      feedback = prepare_coffee_feedback(fields)
-      remaining = fields.reject{|k,v| feedback[:keys].include?(k)}
-      
-      msg = "additional fields:\n" + remaining.map{|k,v| "- #{k}: #{v}"}.sort.join("\n")
-      
-      unless test
-        add_coffee_feedback(opp, feedback[:fields])
-        client.add_note(opp, "Imported coffee feedback; " + msg)
-      else
-        log.log("Ready to add coffee feedback:\n" + feedback[:fields].map{|v| "- #{v['text']}: #{v['value']}"}.sort.join("\n"))
-        log.log(msg)
-      end
-    }
-    
-    log.log("Finished importing from BigQuery table: #{table.query_id}\n" + JSON.pretty_generate(counts))
-  end
-
-  def find_or_create_opportunity(params, tag, test)
-    opp = find_opportunity(params)
-    if opp
-      is_new = false
-      client.add_tag(opp, AUTO_TAG_PREFIX + tag + "-existing") unless test
-    else
-      is_new = true
-      create_params = {
-        name: params[:name],
-        postings: [params[:posting]],
-        origin: ([(params[:origin] || 'sourced').downcase] & ['agency', 'applied', 'internal', 'referred', 'sourced', 'university']).first,
-        sources: Array(params[:source]),
-        tags: [AUTO_TAG_PREFIX + tag + "-new"] + (Array(params[:tags]) || [])
-      }
-      create_params[:emails] = [params[:email]] if !(params[:email] || '').empty?
-      create_params[:links] = [params[:linkedin]] if !(params[:linkedin] || '').empty?
-      create_params[:createdAt] = params[:createdAt] if params[:createdAt]
-      create_params[:stage] = params[:stage_id] if params[:stage_id]
-      create_params[:phones] = Array(params[:phones]) if params[:phones]
-      create_params[:location] = params[:location] if params[:location]
-      create_params[:headline] = params[:headline] if params[:headline]
-      
-      unless test
-        opp = create_opportunity(create_params)
-      else
-        log.log("Ready to create opportunity:\n" + JSON.pretty_generate(create_params))
-      end
-    end
-    [opp, is_new]
-  end
-
-  def find_opportunity(params)
-    opp = nil
-    if params[:email].match?(/.+@.+\..+/)
-      opp = client.opportunities_for_contact(params[:email]).select {|opp|
-        ['', params[:posting]].include?(Util.view_flat(opp)['application__posting'] || '')
-      }.sort_by {|opp|
-        opp['lastInteractionAt']
-      }.last
-    end
-    
-    if opp.nil? && params[:linkedin].match?(/linkedin\.com\/.+/)
-      ids = bigquery.query("SELECT id FROM #{bigquery.table.query_id}_view WHERE LOWER(links) LIKE '%#{Util.escape_sql(params[:linkedin])}%' AND application__posting IN ('#{Util.escape_sql(params[:posting])}', '', null) ORDER BY lastInteractionAt DESC LIMIT 1", '')
-      opp = client.get_opportunity(ids[0][:id]) if ids[0]
-    end
-    
-    log.log("Existing opportunity: #{opp['id']}") if opp
-    opp
-  end
-
-  def create_opportunity(params)
-    opp = client.create_opportunity(params)['data']
-    log.log("Created opportunity: #{opp['id']}") if opp
-    opp
-  end
-
-  def prepare_coffee_feedback(fields)
-    client.prepare_feedback(COFFEE_FEEDBACK_FORM, fields)
-  end
-  
-  def add_coffee_feedback(opp, fields)
-    client.add_profile_form(opp, COFFEE_FEEDBACK_FORM, fields)
-  end
+  end 
 
 end
